requirements.txt
kafka-python==2.2.2
elasticsearch==9.1.0
opensearch-py==3.0.0
requests==2.32.5
python-dateutil==2.9.0.post0

